+++ lib/src/storage/di/column_strategy.dart
@@ -1,10 +1,9 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../domain/repo/persistent_storage.dart' as s;
+import '../domain/entity/column_strategy.dart';
+import '../domain/usecase/column_strategy.dart';
 import 'storage.dart';
 
-part 'column_strategy.freezed.dart';
 part 'column_strategy.g.dart';
 
 @riverpod
@@ -32,52 +31,4 @@ class ColumnStrategy extends _$ColumnStrategy {
     setColumnStrategyUseCase(storage, 2, extent);
     state = ColumnStrategyEnum.minExtent(extent);
   }
-}
-
-@freezed
-sealed class ColumnStrategyEnum with _$ColumnStrategyEnum {
-  const factory ColumnStrategyEnum.fixedCount(final int numChildren) =
-      ColumnStrategyFixedCount;
-
-  const factory ColumnStrategyEnum.maxExtent(final int extent) =
-      ColumnStrategyMaxExtent;
-
-  const factory ColumnStrategyEnum.minExtent(final int extent) =
-      ColumnStrategyMinExtent;
-}
-
-ColumnStrategyEnum initializeColumnStrategyUseCase(
-  final s.PersistentStorage? storage,
-) {
-  if (storage == null) {
-    return const ColumnStrategyEnum.minExtent(440);
-  }
-  final type = storage.getInt('columnStrategyType');
-  final value = storage.getInt('columnStrategyValue');
-  if (type == null || value == null) {
-    return const ColumnStrategyEnum.minExtent(440);
-  }
-  switch (type) {
-    case 0:
-      return ColumnStrategyEnum.fixedCount(value);
-    case 1:
-      return ColumnStrategyEnum.maxExtent(value);
-    case 2:
-      return ColumnStrategyEnum.minExtent(value);
-    default:
-      return const ColumnStrategyEnum.minExtent(440);
-  }
-}
-
-void setColumnStrategyUseCase(
-  final s.PersistentStorage? storage,
-  final int strategy,
-  final int value,
-) {
-  if (storage == null) {
-    return;
-  }
-  storage
-    ..setInt('columnStrategyType', strategy)
-    ..setInt('columnStrategyValue', value);
 }
